Clean up Makefile and complete token type names
- Consolidated duplicate test/all targets in Makefile
- Added test-lexer target for compiler tests
- Completed token_type_to_string for all 51 token types
- Fixed sign-compare warning in lexer keyword matching
- Newline tokens now display inline as '\n'

+++ src/compiler/lexer.c
@@ -190,7 +190,7 @@ static Token* read_identifier(Lexer* lexer) {
     // Check if it's a keyword
     int length = (int)(lexer->current - start);
     for (int i = 0; keywords[i].keyword != NULL; i++) {
-        if (strlen(keywords[i].keyword) == length &&
+        if (strlen(keywords[i].keyword) == (size_t)length &&
             strncmp(keywords[i].keyword, start, length) == 0) {
             return make_token(lexer, keywords[i].type, start);
         }
